+++ test/sandbox_digilocker_sdk_test.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sandbox_digilocker_sdk/sandbox_digilocker_sdk.dart';
 
@@ -7,23 +8,20 @@ void main() {
 
     setUp(() {
       digilockerSDK = DigilockerSDK.instance;
-      digilockerSDK.apiKey = null;
-      digilockerSDK.options = {};
     });
 
-    group('#setAPIKey', () {
-      test('passes API key correctly', () {
-        digilockerSDK.setAPIKey('key_test_1DP5mmOlF5G5aa');
-
-        expect(digilockerSDK.apiKey, equals('key_test_1DP5mmOlF5G5aa'));
-      });
-
-      test('throws error if key does not start with "key_"', () {
-        expect(
-          () => digilockerSDK.setAPIKey('invalid_test_1DP5mmOlF5G5aa'),
-          throwsA(isA<Exception>().having((e) => e.toString(), 'message', contains('API key must start with "key_"'))),
-        );
-      });
+    testWidgets('throws exception if open is called without setting API key', (WidgetTester tester) async {
+      // Create a minimal widget to obtain a BuildContext
+      late BuildContext capturedContext;
+      await tester.pumpWidget(
+        Builder(
+          builder: (context) {
+            capturedContext = context;
+            return const SizedBox.shrink();
+          },
+        ),
+      );
+      expect(() => digilockerSDK.open(context: capturedContext, options: {}), throwsA(isA<Exception>()));
     });
   });
 }