+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## 1.0.0-alpha
 
-* Initial alpha release of the Sandbox DigiLocker SDK.
-* Core functionality for DigiLocker integration.
+* Initial alpha release of the Sandbox Digilocker SDK.
+* Core functionality for Digilocker integration.