+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
-## 1.0.0-alpha
+## 1.0.0
 
-* Initial alpha release of the Sandbox Digilocker SDK.
+* Initial release of the Sandbox Digilocker Flutter SDK.
 * Core functionality for Digilocker integration.